refactor: rename files to use snailcase
