+++ Main/src/init_share.ts
